+++ weaver/processes/wps_process_base.py
@@ -1,9 +1,9 @@
+import os
 from abc import abstractmethod
 from time import sleep
 from typing import TYPE_CHECKING
 
 from pyramid.httpexceptions import HTTPBadGateway
-from pyramid_celery import celery_app as app
 
 from weaver.formats import CONTENT_TYPE_APP_JSON
 from weaver.utils import get_cookie_headers, get_settings, request_extra
@@ -25,7 +25,7 @@ class WpsProcessInterface(object):
                 workflow_inputs,        # type: CWL_RuntimeInputsMap
                 out_dir,                # type: str
                 expected_outputs,       # type: Dict[str, str]
-                ):
+                ):                      # type: (...) -> None
         """
         Execute a remote process using the given inputs.
 
@@ -46,7 +46,7 @@ class WpsProcessInterface(object):
         else:
             self.cookies = {}
         self.headers = {"Accept": CONTENT_TYPE_APP_JSON, "Content-Type": CONTENT_TYPE_APP_JSON}
-        self.settings = get_settings(app)
+        self.settings = get_settings()
 
     def make_request(self, method, url, retry, status_code_mock=None, **kwargs):
         response = request_extra(method, url=url, settings=self.settings,
@@ -59,12 +59,10 @@ class WpsProcessInterface(object):
             response.status_code = status_code_mock
         return response
 
-    @staticmethod
-    def host_file(file_name):
-        settings = get_settings(app)
-        weaver_output_url = get_wps_output_url(settings)
-        weaver_output_dir = get_wps_output_dir(settings)
-        file_name = file_name.replace("file://", "")
+    def host_file(self, file_name):
+        weaver_output_url = get_wps_output_url(self.settings)
+        weaver_output_dir = get_wps_output_dir(self.settings)
+        file_name = os.path.realpath(file_name.replace("file://", ""))  # in case CWL->WPS outputs link was made
 
         if not file_name.startswith(weaver_output_dir):
             raise Exception("Cannot host files outside of the output path : {0}".format(file_name))